Fixed speed variable, comes as float not int from ECU

+++ ClassLibrary/Vehicle.h
@@ -22,8 +22,8 @@ public:
 	uint8_t gear;
 	uint8_t prevGear;
 
-	uint16_t speed;
-	uint16_t prevSpeed;
+	float speed;
+	float prevSpeed;
 
 	float voltage;
 	float prevVoltage;
